Auction update

+++ src/net/connection/Buffer.java
@@ -151,7 +151,7 @@ public class Buffer {
 		writeByte(stuff.getGemSlot3().getValue());
 		writeByte(stuff.getGemBonusType().getValue());
 		writeInt(stuff.getGemBonusValue());
-		writeInt(stuff.getLevel());
+		writeByte(stuff.getLevel());
 		writeByte(stuff.getWear().getValue());
 		writeInt(stuff.getCritical());
 		writeInt(stuff.getStrength());
